refactor(cli): keep Web handoff within lint budgets
Problem:
- The initial Web handoff implementation extended already-baselined CLI and TUI files beyond repository size and complexity budgets.

Root cause:
- Browser lifecycle, frontend startup, and new tests were placed beside legacy monolithic command paths.

Fix:
- Extract Web handoff state, browser readiness, Serve frontend lifecycle, and focused tests into bounded files.
- Preserve reasonix serve and reasonix web behavior without updating the repolint baseline.

Verification:
- go test ./... -count=1
- focused go test -race for CLI and Serve Web paths
- go vet ./internal/cli ./internal/serve
- go run ./tools/repolint
- Windows amd64 cross-compile for internal/cli
- make build
